# vts_server/app/api/api.yaml
swagger: "2.0"
info: 
  title: WebRTC API
  description: Test WebRTC API with express
  version: 1.0 
  contact: 
    email: ykim5470@vreducation.kr
  servers: 
    -url: "https://106.255.237.50:4000/"
    description: Server host url 

schemes: [https]

paths:
  /createRoomNumber: 
    get: 
      summary: Returns a random room number
      responses: 
        200:
          description: Success
          schema:
            type: object
            properties:
              roomId:
                type: string
            example: {roomId: e26a3d4c-276e-4794-a310-d093b15beef3}
        400:
          description: User Error (Invalid Request)
        404: 
          description: Not Found Error 
        500:
          description: Internal Server Error 

  /roomList:
    get:
      summary: Returns a live channel list
      description: Array Return 
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                RoomId:
                  type: string,
                setConfig:
                  type: array
                  items:
                    type: object
                    properties:
                      Title:
                        type: string
                      Host:
                        type: string
                      Thumbnail:
                        type: string
                      RoomCategory:
                        type: string
                      CreatedAt:
                        type: string
            example: [{RoomId: ba11a782-c2bf-403c-9902-1794f5779ecb, setConfig: [{Title: test, Host: host, Thumbnail: uploads/GUIDE/streaming/live/thumbnailSource/xxxx.png, RoomCategory: 'living', CreatedAt: '2022-04-12T04:20:58:000Z'}]}]
        400:
          description: User Error (Invalid Request)

                    
  /guideRoomList:
      get:
        summary: Returns a whole live channel list
        description: Array Return 
        responses:
          200:
            description: Success
            schema:
              type: array
              items:
                type: object
                properties:
                  RoomId:
                    type: string,
                  setConfig:
                    type: array
                    items:
                      type: object
                      properties:
                        Title:
                          type: string
                        Host:
                          type: string
                        Thumbnail:
                          type: string
                        RoomCategory:
                          type: string
                        CreatedAt:
                          type: string
              example: [{RoomId: ba11a782-c2bf-403c-9902-1794f5779ecb, setConfig: [{Title: test, Host: host, Thumbnail: uploads/GUIDE/streaming/live/thumbnailSource/xxxx.png, RoomCategory: 'living', CreatedAt: '2022-04-12T04:20:58:000Z'}]}]
          400:
            description: User Error (Invalid Request)